test cases for inserting XML columns in generated test cases
