+++ core/src/test/kotlin/org/evomaster/core/output/TestCaseWriterTest.kt
@@ -6,12 +6,7 @@ import org.evomaster.core.database.schema.Column
 import org.evomaster.core.database.schema.ColumnDataType.*
 import org.evomaster.core.database.schema.ForeignKey
 import org.evomaster.core.database.schema.Table
-import org.evomaster.core.problem.rest.RestAction
-import org.evomaster.core.problem.rest.RestIndividual
-import org.evomaster.core.problem.rest.SampleType
-import org.evomaster.core.search.ActionResult
-import org.evomaster.core.search.EvaluatedIndividual
-import org.evomaster.core.search.FitnessValue
+import org.evomaster.core.output.EvaluatedIndividualBuilder.Companion.buildEvaluatedIndividual
 import org.evomaster.core.search.gene.*
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
@@ -86,25 +81,6 @@ class TestCaseWriterTest {
     }
 
 
-    private fun buildEvaluatedIndividual(dbInitialization: MutableList<DbAction>): Triple<OutputFormat, String, EvaluatedIndividual<RestIndividual>> {
-        val format = OutputFormat.JAVA_JUNIT_4
-
-        val baseUrlOfSut = "baseUrlOfSut"
-
-        val sampleType = SampleType.RANDOM
-
-        val restActions = emptyList<RestAction>().toMutableList()
-
-
-        val individual = RestIndividual(restActions, sampleType, dbInitialization)
-
-        val fitnessVal = FitnessValue(0.0)
-
-        val results = emptyList<ActionResult>().toMutableList()
-
-        val ei = EvaluatedIndividual<RestIndividual>(fitnessVal, individual, results)
-        return Triple(format, baseUrlOfSut, ei)
-    }
 
     @Test
     fun testTwoAction() {
@@ -758,179 +734,6 @@ class TestCaseWriterTest {
         assertEquals(expectedLines.toString(), lines.toString())
     }
 
-    @Test
-    fun testJSONBEmpty() {
 
-
-        val idColumn = Column("Id", INTEGER, 10, primaryKey = true, autoIncrement = true)
-        val jsonbColumn = Column("jsonbColumn", JSONB, 10, primaryKey = false, autoIncrement = false)
-
-        val table = Table("Table0", setOf(idColumn, jsonbColumn), HashSet<ForeignKey>())
-
-
-        val autoGene = SqlAutoIncrementGene(table.name)
-        val pkGene0 = SqlPrimaryKeyGene(idColumn.name, "Table0", autoGene, 10)
-        val objectGene = ObjectGene(jsonbColumn.name, listOf())
-        val insert = DbAction(table, setOf(idColumn, jsonbColumn), 0L, listOf(pkGene0, objectGene))
-
-        val (format, baseUrlOfSut, ei) = buildEvaluatedIndividual(mutableListOf(insert))
-        val config = EMConfig()
-        config.outputFormat = format
-
-        val test = TestCase(test = ei, name = "test")
-
-        val writer = TestCaseWriter()
-
-        val lines = writer.convertToCompilableTestCode(config, test, baseUrlOfSut)
-
-        val expectedLines = Lines().apply {
-            add("@Test")
-            add("public void test() throws Exception {")
-            indent()
-            add("List<InsertionDto> insertions = sql().insertInto(\"Table0\", 0L)")
-            indent()
-            indent()
-            add(".d(\"jsonbColumn\", \"'{}'\")")
-            deindent()
-            add(".dtos();")
-            deindent()
-            add("controller.execInsertionsIntoDatabase(insertions);")
-            deindent()
-            add("}")
-        }
-
-        assertEquals(expectedLines.toString(), lines.toString())
-    }
-
-
-    @Test
-    fun testJSONBColumnType() {
-
-
-        val idColumn = Column("Id", INTEGER, 10, primaryKey = true, autoIncrement = true)
-        val jsonbColumn = Column("jsonbColumn", JSONB, 10, primaryKey = false, autoIncrement = false)
-
-        val table = Table("Table0", setOf(idColumn, jsonbColumn), HashSet<ForeignKey>())
-
-
-        val autoGene = SqlAutoIncrementGene(table.name)
-        val pkGene0 = SqlPrimaryKeyGene(idColumn.name, "Table0", autoGene, 10)
-        val objectGene = ObjectGene(jsonbColumn.name, listOf(IntegerGene("integerField")))
-        val insert = DbAction(table, setOf(idColumn, jsonbColumn), 0L, listOf(pkGene0, objectGene))
-
-        val (format, baseUrlOfSut, ei) = buildEvaluatedIndividual(mutableListOf(insert))
-        val config = EMConfig()
-        config.outputFormat = format
-
-        val test = TestCase(test = ei, name = "test")
-
-        val writer = TestCaseWriter()
-
-        val lines = writer.convertToCompilableTestCode(config, test, baseUrlOfSut)
-
-        val expectedLines = Lines().apply {
-            add("@Test")
-            add("public void test() throws Exception {")
-            indent()
-            add("List<InsertionDto> insertions = sql().insertInto(\"Table0\", 0L)")
-            indent()
-            indent()
-            add(".d(\"jsonbColumn\", \"'{\\\"integerField\\\":0}'\")")
-            deindent()
-            add(".dtos();")
-            deindent()
-            add("controller.execInsertionsIntoDatabase(insertions);")
-            deindent()
-            add("}")
-        }
-
-        assertEquals(expectedLines.toString(), lines.toString())
-    }
-
-
-    @Test
-    fun testJSONBoolean() {
-        val idColumn = Column("Id", INTEGER, 10, primaryKey = true, autoIncrement = true)
-        val jsonbColumn = Column("jsonbColumn", JSONB, 10, primaryKey = false, autoIncrement = false)
-
-        val table = Table("Table0", setOf(idColumn, jsonbColumn), HashSet<ForeignKey>())
-
-
-        val autoGene = SqlAutoIncrementGene(table.name)
-        val pkGene0 = SqlPrimaryKeyGene(idColumn.name, "Table0", autoGene, 10)
-        val objectGene = ObjectGene(jsonbColumn.name, listOf(BooleanGene("booleanField")))
-        val insert = DbAction(table, setOf(idColumn, jsonbColumn), 0L, listOf(pkGene0, objectGene))
-
-        val (format, baseUrlOfSut, ei) = buildEvaluatedIndividual(mutableListOf(insert))
-        val config = EMConfig()
-        config.outputFormat = format
-
-        val test = TestCase(test = ei, name = "test")
-
-        val writer = TestCaseWriter()
-
-        val lines = writer.convertToCompilableTestCode(config, test, baseUrlOfSut)
-
-        val expectedLines = Lines().apply {
-            add("@Test")
-            add("public void test() throws Exception {")
-            indent()
-            add("List<InsertionDto> insertions = sql().insertInto(\"Table0\", 0L)")
-            indent()
-            indent()
-            add(".d(\"jsonbColumn\", \"'{\\\"booleanField\\\":true}'\")")
-            deindent()
-            add(".dtos();")
-            deindent()
-            add("controller.execInsertionsIntoDatabase(insertions);")
-            deindent()
-            add("}")
-        }
-
-        assertEquals(expectedLines.toString(), lines.toString())
-    }
-
-
-    @Test
-    fun testJSONString() {
-        val idColumn = Column("Id", INTEGER, 10, primaryKey = true, autoIncrement = true)
-        val jsonbColumn = Column("jsonbColumn", JSONB, 10, primaryKey = false, autoIncrement = false)
-
-        val table = Table("Table0", setOf(idColumn, jsonbColumn), HashSet<ForeignKey>())
-
-
-        val autoGene = SqlAutoIncrementGene(table.name)
-        val pkGene0 = SqlPrimaryKeyGene(idColumn.name, "Table0", autoGene, 10)
-        val objectGene = ObjectGene(jsonbColumn.name, listOf(StringGene("stringField")))
-        val insert = DbAction(table, setOf(idColumn, jsonbColumn), 0L, listOf(pkGene0, objectGene))
-
-        val (format, baseUrlOfSut, ei) = buildEvaluatedIndividual(mutableListOf(insert))
-        val config = EMConfig()
-        config.outputFormat = format
-
-        val test = TestCase(test = ei, name = "test")
-
-        val writer = TestCaseWriter()
-
-        val lines = writer.convertToCompilableTestCode(config, test, baseUrlOfSut)
-
-        val expectedLines = Lines().apply {
-            add("@Test")
-            add("public void test() throws Exception {")
-            indent()
-            add("List<InsertionDto> insertions = sql().insertInto(\"Table0\", 0L)")
-            indent()
-            indent()
-            add(".d(\"jsonbColumn\", \"'{\\\"stringField\\\":\\\"foo\\\"}'\")")
-            deindent()
-            add(".dtos();")
-            deindent()
-            add("controller.execInsertionsIntoDatabase(insertions);")
-            deindent()
-            add("}")
-        }
-
-        assertEquals(expectedLines.toString(), lines.toString())
-    }
 
 }
